+++ ReleaseNotes.md
@@ -6,3 +6,9 @@
 
 - Fix for incorrect logging message
 - Changed `IServiceInstance.RemotePath` to `string` to support non-local paths
+
+# 0.3.0
+
+- Introduces tags for instances
+- Introduces support for instance filtering when deploying (using `IServiceFilter`)
+- Introduces new `Deploys(IServiceManager)` extension to simplify Task declarations